+++ backend/src/test/java/com/woowacourse/thankoo/coupon/application/CouponQueryServiceTest.java
@@ -1,0 +1,80 @@
+package com.woowacourse.thankoo.coupon.application;
+
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.MESSAGE;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.NOT_USED;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TITLE;
+import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TYPE;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_EMAIL;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_NAME;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_SOCIAL_ID;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.IMAGE_URL;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_EMAIL;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_NAME;
+import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_SOCIAL_ID;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import com.woowacourse.thankoo.common.annotations.ApplicationTest;
+import com.woowacourse.thankoo.coupon.application.dto.ContentRequest;
+import com.woowacourse.thankoo.coupon.application.dto.CouponRequest;
+import com.woowacourse.thankoo.coupon.domain.CouponRepository;
+import com.woowacourse.thankoo.coupon.presentation.dto.CouponResponse;
+import com.woowacourse.thankoo.member.domain.Member;
+import com.woowacourse.thankoo.member.domain.MemberRepository;
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+@DisplayName("CouponQueryService 는 ")
+@ApplicationTest
+class CouponQueryServiceTest {
+
+    @Autowired
+    private CouponQueryService couponQueryService;
+
+    @Autowired
+    private CouponService couponService;
+
+    @Autowired
+    private CouponRepository couponRepository;
+
+    @Autowired
+    private MemberRepository memberRepository;
+
+    @DisplayName("받은 쿠폰을 조회한다.")
+    @Test
+    void getReceivedCoupons() {
+        Member sender = memberRepository.save(new Member(HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, IMAGE_URL));
+        Member receiver = memberRepository.save(new Member(SKRR_NAME, SKRR_EMAIL, SKRR_SOCIAL_ID, IMAGE_URL));
+
+        couponService.saveAll(sender.getId(), new CouponRequest(List.of(receiver.getId()),
+                new ContentRequest(TYPE, TITLE, MESSAGE)));
+
+        List<CouponResponse> responses = couponQueryService.getReceivedCoupons(receiver.getId(), NOT_USED);
+
+        assertAll(
+                () -> assertThat(responses).hasSize(1),
+                () -> assertThat(responses.get(0).getSender().getId()).isEqualTo(sender.getId()),
+                () -> assertThat(responses.get(0).getReceiver().getId()).isEqualTo(receiver.getId())
+        );
+    }
+
+    @DisplayName("보낸 쿠폰을 조회한다.")
+    @Test
+    void getSentCoupons() {
+        Member sender = memberRepository.save(new Member(HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, IMAGE_URL));
+        Member receiver = memberRepository.save(new Member(SKRR_NAME, SKRR_EMAIL, SKRR_SOCIAL_ID, IMAGE_URL));
+
+        couponService.saveAll(sender.getId(), new CouponRequest(List.of(receiver.getId()),
+                new ContentRequest(TYPE, TITLE, MESSAGE)));
+
+        List<CouponResponse> responses = couponQueryService.getSentCoupons(sender.getId());
+
+        assertAll(
+                () -> assertThat(responses).hasSize(1),
+                () -> assertThat(responses.get(0).getSender().getId()).isEqualTo(sender.getId()),
+                () -> assertThat(responses.get(0).getReceiver().getId()).isEqualTo(receiver.getId())
+        );
+    }
+}
