+++ backend/src/test/java/com/woowacourse/thankoo/coupon/application/CouponServiceTest.java
@@ -1,7 +1,6 @@
 package com.woowacourse.thankoo.coupon.application;
 
 import static com.woowacourse.thankoo.common.fixtures.CouponFixture.MESSAGE;
-import static com.woowacourse.thankoo.common.fixtures.CouponFixture.NOT_USED;
 import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TITLE;
 import static com.woowacourse.thankoo.common.fixtures.CouponFixture.TYPE;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.HUNI_EMAIL;
@@ -16,14 +15,12 @@ import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_NAME;
 import static com.woowacourse.thankoo.common.fixtures.MemberFixture.SKRR_SOCIAL_ID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.woowacourse.thankoo.common.annotations.ApplicationTest;
 import com.woowacourse.thankoo.coupon.application.dto.ContentRequest;
 import com.woowacourse.thankoo.coupon.application.dto.CouponRequest;
 import com.woowacourse.thankoo.coupon.domain.Coupon;
 import com.woowacourse.thankoo.coupon.domain.CouponRepository;
-import com.woowacourse.thankoo.coupon.presentation.dto.CouponResponse;
 import com.woowacourse.thankoo.member.domain.Member;
 import com.woowacourse.thankoo.member.domain.MemberRepository;
 import com.woowacourse.thankoo.member.exception.InvalidMemberException;
@@ -88,41 +85,5 @@ class CouponServiceTest {
                     .isInstanceOf(InvalidMemberException.class)
                     .hasMessage("존재하지 않는 회원입니다.");
         }
-    }
-
-    @DisplayName("받은 쿠폰을 조회한다.")
-    @Test
-    void getReceivedCoupons() {
-        Member sender = memberRepository.save(new Member(HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, IMAGE_URL));
-        Member receiver = memberRepository.save(new Member(SKRR_NAME, SKRR_EMAIL, SKRR_SOCIAL_ID, IMAGE_URL));
-
-        couponService.saveAll(sender.getId(), new CouponRequest(List.of(receiver.getId()),
-                new ContentRequest(TYPE, TITLE, MESSAGE)));
-
-        List<CouponResponse> responses = couponService.getReceivedCoupons(receiver.getId(), NOT_USED);
-
-        assertAll(
-                () -> assertThat(responses).hasSize(1),
-                () -> assertThat(responses.get(0).getSender().getId()).isEqualTo(sender.getId()),
-                () -> assertThat(responses.get(0).getReceiver().getId()).isEqualTo(receiver.getId())
-        );
-    }
-
-    @DisplayName("보낸 쿠폰을 조회한다.")
-    @Test
-    void getSentCoupons() {
-        Member sender = memberRepository.save(new Member(HUNI_NAME, HUNI_EMAIL, HUNI_SOCIAL_ID, IMAGE_URL));
-        Member receiver = memberRepository.save(new Member(SKRR_NAME, SKRR_EMAIL, SKRR_SOCIAL_ID, IMAGE_URL));
-
-        couponService.saveAll(sender.getId(), new CouponRequest(List.of(receiver.getId()),
-                new ContentRequest(TYPE, TITLE, MESSAGE)));
-
-        List<CouponResponse> responses = couponService.getSentCoupons(sender.getId());
-
-        assertAll(
-                () -> assertThat(responses).hasSize(1),
-                () -> assertThat(responses.get(0).getSender().getId()).isEqualTo(sender.getId()),
-                () -> assertThat(responses.get(0).getReceiver().getId()).isEqualTo(receiver.getId())
-        );
     }
 }
